requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.7.0
anthropic==0.42.0
openpyxl==3.1.5
python-multipart==0.0.18
jinja2==3.1.4
cachetools==5.5.0
python-dotenv==1.0.1
aiofiles==24.1.0
